backend/requirements.txt
aiofiles==25.1.0
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
anywidget==0.9.21
APScheduler==3.11.2
asttokens==3.0.1
beautifulsoup4==4.14.3
billiard==4.2.4
celery==5.6.2
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
comm==0.2.3
contourpy==1.3.3
curl_cffi==0.13.0
cycler==0.12.1
dateparser==1.3.0
decorator==5.2.1
dill==0.4.1
executing==2.2.1
fastapi==0.129.0
filelock==3.24.3
finnhub-python==2.4.27
fonttools==4.61.1
frozendict==2.4.7
fsspec==2026.2.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
ipython==9.10.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.8
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
kombu==5.6.2
llvmlite==0.44.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.12
mypy_extensions==1.1.0
narwhals==2.16.0
networkx==3.6.1
numba==0.61.2
numpy==2.2.6
packaging==26.0
pandas==2.3.3
pandas-ta==0.4.71b0
parso==0.8.6
peewee==3.19.0
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
plotly==6.5.2
prompt_toolkit==3.0.52
protobuf==6.33.5
psygnal==0.15.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
redis==7.2.0
regex==2026.2.19
requests==2.32.5
rich==14.3.3
safetensors==0.7.0
schedule==1.2.2
scikit-learn==1.8.0
scipy==1.17.0
sentencepiece==0.2.1
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
traitlets==5.14.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vectorbt==0.28.4
vine==5.1.0
wcwidth==0.6.0
websockets==16.0
widgetsnbextension==4.0.15
xgboost==3.2.0
yfinance==1.2.0
